refactor collators table
